Integration test and small changes

+++ src/main/java/com/iozeta/SpringIOZeta/Database/Repositories/ProgressRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface ProgressRepository extends JpaRepository<Progress, Integer> {
+public interface ProgressRepository extends JpaRepository<Progress, Long> {
     List<Progress> findAllBy();
 }
